Fix permissions bug in IRC Rename Permissions class to remove confusion with the Facade
